Add aggress.rb (it's now the light-blue in play_one_game.sh)
It's a fork of explore.rb with an improved algorithm I wrote
mostly based on Sam and I's discussions.

(The only things I used that I don't remember us discussing
are the sub-linear functions I combined different boringnesses with,
and the 0.1 tolerance when randomly choosing the best space to go to.
These seemed to work well, so I used them.)

It beats explore.rb handily.

It still does not do as much with sight radius information
as it could.